Refactoring IO (n+2) Finished the remaining chunk readers.

+++ src/main/kotlin/spirite/base/file/v2/converters/SifWorkspaceExporter.kt
@@ -350,7 +350,7 @@ class SifWorkspaceExporter(
                 SifViewView.Properties(
                     if(nodeProps.isVisible) 1 else 0,
                     nodeProps.alpha,
-                    nodeProps.method.methodType.ordinal.toByte(),
+                    nodeProps.method.methodType.ordinal,
                     nodeProps.method.renderValue,
                     nodeProps.ox,
                     nodeProps.oy)
