+++ src/main/kotlin/spirite/core/file/contracts/SifViewChunk.kt
@@ -9,7 +9,7 @@ class SifViewView(
     class Properties(
         val bitmap: Byte,
         val alpha: Float,
-        val renderMethod: Byte,
+        val renderMethod: Int,
         val renderValue: Int,
         val ox: Int,
         val oy: Int)
